Move notifications file into util dir
