+++ src/web/containers/__engineering__/PIDGainsPanel/PIDGainsPanel.tsx
@@ -4,7 +4,7 @@
 /* eslint-disable no-unused-vars */
 
 import React from "react";
-import { error, success, warning, info, debug } from "../../../notifications/notifications";
+import { error, success, warning, info, debug } from "../../../utils/notifications";
 import Button from "@mui/material/Button";
 import {
     BotStatus,
